+++ cmake/Modules/BuildFLTK.cmake
@@ -9,8 +9,8 @@ set(FLTK_GIT_REPOSITORY "https://github.com/fltk/fltk.git")
 # The cutting EDGE!
 #set( FLTK_GIT_TAG master )
 
-set(FLTK_GIT_TAG 81a5736006d64d749f9694268dd180ae80ec3d9d)
-#set(FLTK_GIT_TAG 27d991f046bdebb12bfd58f7c05a19f135979c29)
+#set(FLTK_GIT_TAG 81a5736006d64d749f9694268dd180ae80ec3d9d) # OK
+set(FLTK_GIT_TAG 27d991f046bdebb12bfd58f7c05a19f135979c29)
 
 if(MRV2_PYFLTK OR FLTK_BUILD_SHARED)
     # If we are building pyFLTK compile shared
@@ -39,6 +39,8 @@ endif()
 # These two are always built by tlRender
 set(FLTK_USE_SYSTEM_ZLIB TRUE)
 set(FLTK_USE_SYSTEM_LIBPNG TRUE)
+
+# Set FLTK default dependencies
 set(FLTK_DEPENDENCIES PNG ZLIB)
 
 # We set this to use FLTK's built-in libdecor