+++ src/lib/mrvCore/mrvWait.h
@@ -1,0 +1,10 @@
+
+#pragma once
+
+namespace mrv
+{
+    namespace wait
+    {
+        void milliseconds(float milliseconds);
+    }
+}