+++ src/lib/mrvFl/mrvCallbacks.cpp
@@ -1685,6 +1685,8 @@ namespace mrv
         if (has_about_window)
             ui->uiAbout->uiMain->show();
 
+        ui->uiView->frameView();
+        
         PanelGroup::show_all();
     }