+++ src/lib/mrvGL/mrvGLWindow.cpp
@@ -11,6 +11,11 @@
 #endif
 
 #ifdef FLTK_USE_WAYLAND
+
+#    include <wayland-client.h>
+
+// Since we cannot include glx.h and egl together, we define some constants
+// and extern EGL functions here.
 extern "C"
 {
     typedef unsigned int EGLBoolean;
@@ -95,5 +100,17 @@ namespace mrv
 #    endif
     }
 #endif
+
+    void GLWindow::show()
+    {
+        Fl_Gl_Window::show();
+        
+// #ifdef FLTK_USE_WAYLAND
+//         if (fl_wl_display())
+//         {
+//             wl_surface_set_opaque_region(fl_wl_surface(fl_wl_xid(this)), NULL);
+//         }
+// #endif
+    }
 
 } // namespace mrv